+++ packages/rocketh-deploy/src/index.ts
@@ -144,7 +144,7 @@ type FactoryParams = {
 	maxPriorityFeePerGas: `0x${string}` | undefined;
 };
 async function getCreate2Factory(env: Environment, signer: Signer, params: FactoryParams) {
-	const deploymentInfo = env.config.network.deterministicDeployment.create2;
+	const deploymentInfo = env.config.target.deterministicDeployment.create2;
 	if (!deploymentInfo) throw new Error('create2 deterministic deployment info not found');
 
 	const factoryAddress = deploymentInfo.factory;
@@ -206,7 +206,7 @@ async function getCreate2Factory(env: Environment, signer: Signer, params: Facto
 }
 
 async function getCreate3Factory(env: Environment, signer: Signer, params: FactoryParams) {
-	const deploymentInfo = env.config.network.deterministicDeployment.create3;
+	const deploymentInfo = env.config.target.deterministicDeployment.create3;
 	if (!deploymentInfo) throw new Error('create3 deterministic deployment info not found');
 
 	const factoryAddress = deploymentInfo.factory;