+++ packages/rocketh-doc/src/index.ts
@@ -38,9 +38,9 @@ function filter(options: RunOptions, name: string): boolean {
 }
 
 export async function run(config: ResolvedConfig, options: RunOptions) {
-	const {deployments, chainId} = loadDeployments(config.deployments, config.network.name);
+	const {deployments, chainId} = loadDeployments(config.deployments, config.target.name);
 	if (!chainId) {
-		throw new Error(`no chainId found for ${config.network.name}`);
+		throw new Error(`no chainId found for ${config.target.name}`);
 	}
 	generate({deployments}, options);
 }