+++ packages/rocketh-verifier/src/metadata.ts
@@ -3,19 +3,19 @@ import fs from 'fs-extra';
 import path from 'path';
 
 export function exportMetadata(config: ResolvedConfig, {out}: {out: string}) {
-	const {deployments, chainId} = loadDeployments(config.deployments, config.network.name, false);
+	const {deployments, chainId} = loadDeployments(config.deployments, config.target.name, false);
 
 	if (Object.keys(deployments).length === 0) {
-		console.log(`the network ${config.network.name} has zero deployments`);
+		console.log(`the target ${config.target.name} has zero deployments`);
 		process.exit();
 	}
 
 	if (!chainId) {
-		console.error(`the network ${config.network.name} has no chainId recorded`);
+		console.error(`the target ${config.target.name} has no chainId recorded`);
 		process.exit(1);
 	}
 
-	const folder = path.join(out, config.network.name);
+	const folder = path.join(out, config.target.name);
 	fs.emptyDirSync(folder);
 	const deploymentNames = Object.keys(deployments);
 	for (const deploymentName of deploymentNames) {