+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocketh/doc
 
+## 1.0.0-testing.0
+
+### Minor Changes
+
+- revamp the settings to be allowed to configure per chain as well as per target
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.15.0-testing.0
+
 ## 0.14.9
 
 ### Patch Changes